Made wip game display ui alongside file parcing
Time: Cannot Be Determined (Im on a plane as of writing this
Moved the game selector bar a bit to the left to make more space for game summary
Game summary ui has: Game name, Game logo, Creator name, Creator Icon, Time spent on game, Promo image (can support gif), game description, and complexity meter.

This was actually supposed to be 2 commits but guess who forgot to hit the button so heres the rest of the info:

The game will properly parse folders found inside of the userdata and turn those into buttons. all the icons work, alongside descriptions. It is only a matter of changing the translations for concat text